cast_rayの関数が多かったので分割して、新しいディレクトリに配置した。
