+++ srcs/game/cast_ray/set_wall_hit_coodinate.c
@@ -1,77 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   cast_ray.c                                         :+:      :+:    :+:   */
+/*   set_wall_hit_coodinate.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: rnakai <rnakai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/11/11 15:24:22 by rnakai            #+#    #+#             */
-/*   Updated: 2020/11/12 22:39:02 by rnakai           ###   ########.fr       */
+/*   Created: 2020/11/12 22:51:57 by rnakai            #+#    #+#             */
+/*   Updated: 2020/11/12 22:53:03 by rnakai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../definitions.h"
-#include <float.h>
-
-float		normalize_angle(float ray_angle)
-{
-	ray_angle = remainder(ray_angle, TWO_PI);
-	if (ray_angle < 0)
-	{
-		ray_angle = TWO_PI + ray_angle;
-	}
-	return (ray_angle);
-}
-
-float		distance_between_points(float x1, float y1, float x2, float y2)
-{
-	return (sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
-}
-
-void		cast_ray(float ray_angle, int strip_id)
-{
-	t_cast_ray_var			hrz;
-	t_cast_ray_var			vrt;
-	t_cast_ray_var_common	cmn;
-
-	ray_angle = normalize_angle(ray_angle);
-	cmn.ray_angle = ray_angle;
-	cmn.strip_id = strip_id;
-	set_where_ray_is_facing(&cmn);
-	//
-	hrz_set_first_intercept(&hrz, &cmn);
-	hrz_cast_ray_until_wall(&hrz, &cmn);
-	//
-	vrt_set_first_intercept(&vrt, &cmn);
-	vrt_cast_ray_until_wall(&vrt, &cmn);
-	//
-	set_ray_distances(&hrz, &vrt);
-	//
-	set_g_rays_each_element(&hrz, &vrt, &cmn);
-}
-
-void		set_where_ray_is_facing(t_cast_ray_var_common *cmn)
-{
-	cmn->is_ray_facing_down =
-		(cmn->ray_angle > 0 && cmn->ray_angle < PI);
-	cmn->is_ray_facing_up = (!cmn->is_ray_facing_down);
-	cmn->is_ray_facing_right =
-		(cmn->ray_angle < 0.5 * PI || cmn->ray_angle > 1.5 * PI);
-	cmn->is_ray_facing_left = (!cmn->is_ray_facing_right);
-}
-
-void		set_ray_distances(t_cast_ray_var *hrz,
-				t_cast_ray_var *vrt)
-{
-	hrz->hit_distance = (hrz->found_wall_hit)
-		? distance_between_points(
-			g_player.x, g_player.y, hrz->wall_hit_x, hrz->wall_hit_y)
-		: FLT_MAX;
-	vrt->hit_distance = (vrt->found_wall_hit)
-		? distance_between_points(
-			g_player.x, g_player.y, vrt->wall_hit_x, vrt->wall_hit_y)
-		: FLT_MAX;
-}
 
 /*
 ** ///////////////////////////////////////////
@@ -197,30 +136,4 @@ void		vrt_cast_ray_until_wall(t_cast_ray_var *vrt,
 			vrt->next_touch_y += vrt->ystep;
 		}
 	}
-}
-
-void		set_g_rays_each_element(t_cast_ray_var *hrz,
-				t_cast_ray_var *vrt, t_cast_ray_var_common *cmn)
-{
-	if (vrt->hit_distance < hrz->hit_distance)
-	{
-		g_rays[cmn->strip_id].distance = vrt->hit_distance;
-		g_rays[cmn->strip_id].wall_hit_x = vrt->wall_hit_x;
-		g_rays[cmn->strip_id].wall_hit_y = vrt->wall_hit_y;
-		g_rays[cmn->strip_id].wall_hit_content = vrt->wall_content;
-		g_rays[cmn->strip_id].was_hit_vertical = TRUE;
-	}
-	else
-	{
-		g_rays[cmn->strip_id].distance = hrz->hit_distance;
-		g_rays[cmn->strip_id].wall_hit_x = hrz->wall_hit_x;
-		g_rays[cmn->strip_id].wall_hit_y = hrz->wall_hit_y;
-		g_rays[cmn->strip_id].wall_hit_content = hrz->wall_content;
-		g_rays[cmn->strip_id].was_hit_vertical = FALSE;
-	}
-	g_rays[cmn->strip_id].ray_angle = cmn->ray_angle;
-	g_rays[cmn->strip_id].is_ray_facing_down = cmn->is_ray_facing_down;
-	g_rays[cmn->strip_id].is_ray_facing_up = cmn->is_ray_facing_up;
-	g_rays[cmn->strip_id].is_ray_facing_left = cmn->is_ray_facing_left;
-	g_rays[cmn->strip_id].is_ray_facing_right = cmn->is_ray_facing_right;
 }
